pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "testfix"
version = "0.1.0"
description = "AI-powered CLI to automatically fix failing tests — non-interactive, pipeable, CI-ready"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "BSL-1.1" }
authors = [{ name = "faw21" }]
keywords = ["ai", "testing", "cli", "developer-tools", "automation", "pytest", "tdd"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Environment :: Console",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Testing",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.0",
    "anthropic>=0.25",
    "openai>=1.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "pytest-mock",
    "anthropic>=0.25",
    "openai>=1.0",
]

[project.scripts]
testfix = "testfix.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/testfix"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=testfix --cov-report=term-missing --cov-fail-under=80"

[tool.coverage.run]
source = ["src/testfix"]
